gdb-cross-canadian: use correct exec-prefix path for python
Incorrect exec-prefix path was given to gdb which leads to gdb
startup failure when SDK is not installed to its original destination.
Gdb relocates the exec-prefix path, so it will work for SDKs that
are installed to different location. PYTHONHOME env in no longer
neeeded for gdb.

Upstreamed to oe-core, putting here until it gets merged

Change-Id: If3f5125a6063bf9eae56b2b6ecbdebb9939ad362
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>