Remove irda, pcmcia, pci, 3g, nfc, zeroconf features from distro
irda is support for infra-red communication
pcmcia and pci are not relevant for any of the reference devices
3g adds Ofono to the stack and since we are not making phones,
this is completely unnecessary waste of resources.
zeroconf adds the Avahi daemon that is a rarely necessary
zero-configuration networking manager.
nfc is something we support, but none of the reference devices
comes with this hardware, and would require an external device that
has not been tested by us at all.  It's likely that an end user
that wants to use custom hardware will need to rebuild the stack
anyway so they can just add any of these additional features back
to their distro as necessary when they build their own stack.

Change-Id: If3951a4170d94ed8db1d40c2bb2d47cf569dc36a
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>