Add Silica ArchiTech tibidabo board
Builds & runs ok!

The current implementation is done with the strategy of using the
vendor's defined repo versions for the yocto stack, instead of the
common versions defined by b2qt. The intention is to build upon the
vendor stack as closely as possible, as that is (hopefully) tested and
maintained.

Task-number: QTEE-801
Change-Id: Ib93b54178de870b5f288b2fae982ea2b4bb74e50
Reviewed-by: Samuli Piippo <samuli.piippo@theqtcompany.com>