Enable to use sources from Qt SDK
This enables sources to be fetched from the Qt SDK, if available, instead of
internal git repositories. Required for BYOS customers to build B2Qt things.
Git repositories in SRC_URI are configured with 'local-uri' parameter, that
defines the path to the source code directory inside the Qt SDK. The base
path to Qt SDK is set in local.conf when setup-environment.sh in run.

Change-Id: I7f5e96fb6b9e9e55cac9d50cefb17830618193b8
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>
Reviewed-by: Andy Nichols <andy.nichols@digia.com>