Export /usr/include/HAL into the toolchain
libgal-dev only provides libGAL.so in the toolchain. That is not enough
when compiling a hardware compositor (e.g. the Weston hw compositor patch)
because the headers are needed too. For Weston having the headers in the
sysroot is enough, but in the future we may need these for QtWayland which
is built with just the toolchain, so include them in the toolchain too.

Change-Id: I81c4609d69da788460071b441cdec26f0dc1cbd0
Reviewed-by: Samuli Piippo <samuli.piippo@digia.com>