nitrogen6x: update kernel configs
Previous kernel config changes are now merged upstream, and the fix
for the imx-vpu is also in, so all configuration changes can be removed.

Add support for Nitrogen6_Lite. The Nitrogen6_Lite board supports
a Broadcom 4331 WiFi/Bluetooth module, and typically has one of its'
GPIO ports configured as an LED_GPIO control (for a dry-contact output).

Change-Id: I840e7703624c652acf0a0262102e6491a3908719
Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>