Performance analysis tools for Linux
Build perf with DWARF unwinding support. Kernels older then 3.14
need to be patched to add ARM support. Kernels for tibidabo (3.0)
and imx53 (2.6) are too old and would require more patches to be
backported, so they are left without unwinding support.

RPi needed extra layer to modify the kernel recipe, and linux-imx
recipe was renamed so that it doesn't conflict with older tibidabo
kernel recipe.

Change-Id: Ic8331f8feeace67b32a5ce99ffd3fda517d595d1
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>
Reviewed-by: Ulf Hermann <ulf.hermann@theqtcompany.com>