Toolchain: add static device mkspec, suitable for Qt Creator
The linux-oe-g++ mkspec used by meta-qt5 depends on the user sourcing
the toolchain environment-setup script, which modifies the PATH and
sets a long range of environment variables. That is not a "best
practice" way of use Qt, and does not allow smooth integration of the
toolchain into Qt Creator.

Here, we instead create a static mkspec for the toolchain:
devices/linux-oe-generic-g++. The toolchain recipe will will the
qdevice.pri file with the required compiler flags etc. for the target.

Additionally, we introduce an extra .pri file for device-specific
settings, so they can be shared between the two mkspecs.

Change-Id: Ie5a019524ba6cdddffb71b0c7c28861ddba17fbd
Reviewed-by: Samuli Piippo <samuli.piippo@theqtcompany.com>