Include Qt recipes only when meta-qt5 layer is used
If the meta-qt5 layer is not included in the BBLAYERS, then do not
include our additional qt recipes from recipes-qt. This way meta-b2qt
can be used with meta-qt5 just to build the base images/toolchain.

By default, all bblayers.conf templates include meta-qt5, so it needs
to be remove first to use this feature.

Change-Id: I73b8f0e65a6f4f950498097a4801ba5f5325cd7d
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@theqtcompany.com>