Add touch_frame events to Weston 1.5.0
Otherwise no Qt applications react on touch. This issue was
eventually resolved upstream by dropping the built-in evdev backend
and relying on libinput exclusively, where it is fixed. The libinput
backend is required since Weston 1.7. In 1.6 the libinput backend
was made the default, but the old backend was still there. Thus this
fix also applies to Weston 1.6.

Task-number: QTEE-831
Change-Id: Ib089136723b7b1dc6f760b5eedb12d7b8a50abbc
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>