Use SSTATE_MIRRORS to speedup the local builds
SSTATE_MIRRORS contains the build artifacts from previous builds
done in CI and Jenkins. When running yocto builds locally for components
with identical build configuration, the artifacts can be fetched from
the mirror. This is usually the case for most of the base components
and toolchain, as they are updated less frequently.

Change-Id: Ic185b5a5dfd3b3b839bcd0ae102067a40a3040c6
Reviewed-by: Teemu Holappa <teemu.holappa@theqtcompany.com>