Add Support for Renesas R-Car Gen2 Devices
I have only personally tested the porter build as it's the only
device I have, but getting the others to work as well should be
straight forward.

Please see the note in conf/distro/include/rcar-gen2.inc about
the binary drivers. The documentation we ship with Qt Creator
should expand on how to install the drivers. Additionally we
must check with Renesas how we could redistribute prebuilt
images if we want to do that.

Change-Id: I82e03aad9b43bee068aaa0472aeac4d394f049a8
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>