intel-corei7-64: fix race in image generation
Prepending to IMAGE_CMD_ext3 in order to push files into the rootfs is
flaky for a couple of reasons. First, it is conceptually inconsistent
with the rest of the build system - files that go into the rootfs should
really by installed by packages through the regular mechanisms. Second,
and what prompted this patch, with several concurrent image generation
tasks this is just unsafe. I regularly experienced .tgz image generation
failures, as tar complained that the files changed from under it.

Since this is related to the grub-efi packaging on this particular device,
just package the microcode into the required initramfs in that recipe.

Change-Id: I4e037ddbb017d1764852310b0b77900cde260dd6
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>