Correctly use QtQuickCompiler when available
Make sure the we use QtQuickCompiler only when we either have access
to the sources (internal gerrit repo) or the prebuilt binaries are available.

The internal-build class checks if we are working within The Qt Company network,
so we will also have access to the sources. Setting ENABLE_QTQUICKCOMPILER env
to "1" in local.conf has the same effect. Otherwise check if the prebuilt
binaries are available as installed by the Qt Online SDK.

Change-Id: I63b09ebabc96bb14b119ea7472a3fa7b10e5c2a8
Task-number: QTBUG-53052
Task-number: QTBUG-53216
Reviewed-by: Teemu Holappa <teemu.holappa@theqtcompany.com>